Soften the requirements of unlink2

+++ R/unlink2.R
@@ -5,11 +5,10 @@
 #'
 #' @details
 #' This variant hardens file path handling, for better Windows compatibility.
+#' Paths that do not exist on disk are silently ignored.
 #'
 #' @param x `character`.
 #' Files or directories to be deleted.
-#' Unlink base `unlink`, these must exist on disk or the function will
-#' intentionally error.
 #'
 #' @return `integer(1)`.
 #' `0` for success, `1` for failure, invisibly.
@@ -23,10 +22,14 @@
 #' out <- unlink2(c(x, y))
 #' print(out)
 unlink2 <- function(x) {
+    x <- x[file.exists(x)]
+    if (!hasLength(x)) {
+        return(invisible(0L))
+    }
     x <- normalizePath(
         path = x,
         winslash = "\\",
-        mustWork = TRUE
+        mustWork = FALSE
     )
     out <- unlink(x, recursive = TRUE)
     assert(!any(file.exists(x)))
